+++ tests/bootstrap.php
@@ -12,3 +12,4 @@
 
 require_once dirname( __DIR__ ) . '/core/QT_Catalog.php';
 require_once dirname( __DIR__ ) . '/core/QT_Prerequisite.php';
+require_once dirname( __DIR__ ) . '/core/QT_Person.php';